Support case ID: 03386738 https://services.arm.com/support/s/case/5004L00000OncY9QAJ/using-ethosu-driver-in-rtos-environment
In the FreeRTOS application example in ethos-u-core-platform the overload of
int ethosu_semaphore_give(void *sem)
can receive an error code from xSemaphoreGive() (queue is full)
which is suppressed (// do nothing) in the example code.

The root cause is an asymetric use of ethosu_semaphore_take/ethosu_semaphore_give in the driver:
- ethosu_release_driver() always gives the semaphore
- ethosu_reserve_driver() takes the semaphore only if no driver is available
In the case of a single task using the driver this leads to constantly giving the semaphore without taking it.

This PR fixes this behaviour by making the semaphore usage symmetric.
As a follow-up, suppressing of the return code could be removed in the application example.

Signed-off-by: Manuel Loew Manuel.Loew@infineon.com
Change-Id: Ib62c974bbc13ae6badf9492294d60508a99cfb4f